布局管理发布成功。依赖布局中不允许有第三方以libs .so方法引入

+++ mylibrary/src/main/java/com/rt/mylibrary/utils/Utils.java
@@ -22,9 +22,9 @@ import java.util.Hashtable;
 
 public class Utils {
 
-    public static Bitmap createQRCode(String url, int width, int height) {
-        int w = width;
-        int h = height;
+    public static Bitmap createQRCode(String url, float width, float height) {
+        int w = (int) width;
+        int h = (int) height;
         try {
             //判断URL合法性
             if (url == null || "".equals(url) || url.length() < 1) {
